Bug 2026297.  a=diannaS DONTBUILD
Original Revision: https://phabricator.services.mozilla.com/D290596

Differential Revision: https://phabricator.services.mozilla.com/D290953

+++ gfx/2d/FilterNodeSoftware.cpp
@@ -2678,9 +2678,11 @@ already_AddRefed<DataSourceSurface> FilterNodeDisplacementMapSoftware::Render(
       uint32_t mapIndex = y * mapStride + 4 * x;
       uint32_t targIndex = y * targetStride + 4 * x;
       int32_t sourceX =
-          x + scaleOver255 * mapData[mapIndex + xChannel] + scaleAdjustment;
+          x + int32_t(scaleOver255 * mapData[mapIndex + xChannel] +
+                      scaleAdjustment);
       int32_t sourceY =
-          y + scaleOver255 * mapData[mapIndex + yChannel] + scaleAdjustment;
+          y + int32_t(scaleOver255 * mapData[mapIndex + yChannel] +
+                      scaleAdjustment);
       *(uint32_t*)(targetData + targIndex) = ColorAtPoint(
           sourceData, sourceStride, sourceBegin, sourceEnd, sourceX, sourceY);
     }
